% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessData.R
\name{ProcessData}
\alias{ProcessData}
\title{Process XML annonated data files}
\usage{
ProcessData(directory)
}
\arguments{
\item{directory}{file path to directory containing individual annonated sleep data files in standard profusion XML format}
}
\value{
Returns a dataframe of sleep summary metrics per input file
}
\description{
A function that takes a directory of xml files and outputs sleep summary metrics
}
\examples{
sleepSummary <- ProcessData("data")

}
