% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runWebApp.R
\name{runWebApp}
\alias{runWebApp}
\title{Launch Shiny App for TestingPackage}
\usage{
runWebApp()
}
\value{
No return value but open up a Shiny page.
}
\description{
A function that launches the Shiny app for rSleepPSGcohort
}
\examples{
\dontrun{

rSleepPSGcohort::runWebApp()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}
}
