% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizeSummaryMetrics.R
\name{VisualizeSummaryMetrics}
\alias{VisualizeSummaryMetrics}
\title{Process XML annonated data files}
\usage{
VisualizeSummaryMetrics(summaryMetrics)
}
\arguments{
\item{summaryMetrics}{dataframe output}
}
\value{
Returns a ggplot bar graph of timeasleep
}
\description{
A function that takes the output from ProcessData and visualizes total sleep time in minutes by participant
}
\examples{
metricsDf <- ProcessData("data")
print(VisualizeSummaryMetrics(metricsDf))

}
